Remove Search Autocomplete
Porting to
https://github.com/workarea-commerce/workarea-classic-search-autocomplete
