--- core/app/assets/javascripts/workarea/core/templates/ui_menu_item.jst.ejs
@@ -1,10 +1,0 @@
-<%
-throw new Error(
-    'workarea/core/templates/ui_menu_item.jst.ejs: this template has ' +
-    'moved from core into both the admin and storefront, respectively. ' +
-    'Please adjust your overridden admin and/or storefront manifests to \n\n' +
-    'Find: "workarea/core/templates/ui_menu_item"\n' +
-    'Replace with either: "workarea/admin/templates/ui_menu_item" or\n' +
-    '                     "workarea/storefront/templates/ui_menu_item"'
-);
-%>
