+++ docs/source/upgrade-guides/from-3-4-to-3-5.html.md
@@ -103,3 +103,13 @@ As a part of this process, the `digital?` flag on products was no longer necessa
 ### What Do You Need To Do?
 
 The v3.5 migration script creates `Fulfillment::Sku` records for existing digital items, setting their policy to `ignore_digital`. This should preserve any behavior in your system for digital products. However, we strongly recommend reworking any code that relies on the `digital?` method of either `Catalog::Product` or `Order::Item`. Instead use `requires_shipping?` on `Order` and `Order::Item` to determine behaviors within this concern.
+
+## Move Storefront Search Autocomplete Functionality to Plugin
+
+### What's Changing?
+
+We've removed the default Search Autocomplete functionality from the Storefront into a plugin to be able to give applications more flexibility.
+
+### What Do You Need To Do?
+
+To keep this functionality present in your application you'll need to install the `workarea-classic_search_autocomplete` plugin.
