--- storefront/app/assets/javascripts/workarea/storefront/templates/ui_menu_heading.jst.ejs
@@ -1,1 +1,0 @@
-<li class="ui-menu-heading"><%= categoryName %></li>
