--- storefront/app/assets/javascripts/workarea/storefront/templates/ui_menu_item.jst.ejs
@@ -1,8 +1,0 @@
-<li class="ui-menu-item">
-    <span class="ui-menu-item-wrapper">
-        <% if (image) { %>
-            <img src="<%= image %>" alt="">
-        <% } %>
-        <%= label %>
-    </span>
-</li>
