--- storefront/app/assets/stylesheets/jquery_ui/storefront/_ui_autocomplete.scss
@@ -1,17 +1,0 @@
-/*------------------------------------*\
-    #UI-AUTOCOMPLETE
-\*------------------------------------*/
-
-$ui-autocomplete-bg-color:  $white !default;
-
-$ui-autocomplete-width:     190px !default;
-
-
-.ui-autocomplete {
-    @extend %list-reset;
-    position: absolute;
-    z-index: index($components, ui-autocomplete);
-    padding: $spacing-unit;
-    width: $ui-autocomplete-width;
-    background: $ui-autocomplete-bg-color;
-}
