+++ storefront/app/assets/stylesheets/workarea/storefront/application.scss.erb
@@ -45,10 +45,8 @@
 <%= append_stylesheets('storefront.typography') %>
 
 @import 'avalanche';
-@import 'jquery_ui/storefront/ui_autocomplete';
 @import 'jquery_ui/storefront/ui_dialog';
 @import 'jquery_ui/storefront/ui_helper_hidden_accessible';
-@import 'jquery_ui/storefront/ui_menu';
 @import 'jquery_ui/storefront/ui_state_focus';
 @import 'jquery_ui/storefront/ui_widget_overlay';
 <%= append_stylesheets('storefront.dependencies') %>
