--- storefront/app/views/workarea/storefront/searches/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.results @results
